backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
paddleocr==2.7.0.3
paddlepaddle==2.6.0
opencv-python==4.9.0.80
pillow==10.2.0
pymupdf==1.23.8
groq==0.4.2
python-dotenv==1.0.0
